Some updates

+++ src/TDNPGL.Networking/Delegates.cs
@@ -4,5 +4,5 @@ namespace TDNPGL.Networking
 {
     public delegate void ClientDisconnect(IChannel channel, byte[] packet);
     public delegate void ClientConnect(IChannel channel);
-    public delegate int ClientPing(IChannel channel, byte[] packet);
+    public delegate byte ClientPing(IChannel channel, byte[] packet);
 }
